src/agartha gateopener/src/requirements.txt
playwright
feedparser
fake-useragent
langchain
langchain-ollama
langgraph
discord.py
requests
sentence-transformers
faiss-cpu
langchain-community
langchain-huggingface
